feat: add button to open file management dialog in duplicate registration warning (#354)
