+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## Not Released Yet
 
 ### 🚀新機能
+- 重複登録の警告にファイル管理ダイアログを開くボタンを追加
 
 ### 🐛バグ修正
 - ログの横スクロールバーが見切れる問題を修正
